requirements.txt
# Earnings Event Trading Agent (EETA) - Dependencies
# Core Data Science
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning / Deep Learning
torch>=2.0.0
scikit-learn>=1.3.0

# Financial Data
yfinance>=0.2.28
finnhub-python>=2.4.18

# Reinforcement Learning Environment
gymnasium>=0.29.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Web Application
streamlit>=1.28.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Data Storage
pyarrow>=13.0.0

# Progress Bars
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
requests>=2.31.0